requirements.txt
Flask
Flask-Session
cs50
werkzeug
openai
python-dotenv
google-cloud-aiplatform
google-generativeai